// src/pages/client/data.d.ts
export type StatisticsDataType = {
  robotCnt: number;
  totalRunHours: number;
  totalRunCount: number;
  totalRunWorkDays: number;
}

export type EnterpriseInfoType = {
  name: string;
  enterpriseDescription: string;
  enterpriseIcon: string;
  principalUserName: string;
  principalUserPhone: string;
}